fix percentage

+++ src/main/java/com/wickend/payload/Receipt.java
@@ -12,5 +12,5 @@ public class Receipt {
     private String presName;
     private String courier;
     private List<Product> products;
-    private Integer percent;
+    private double percent;
 }
